// react-web-app/src/models/brewery/Brewery.tsx
export class Brewery {

    breweryId!: string;
    breweryName!: string;
    breweryType?: string;
    address!: string;
    city!: string;
    stateProvince!: string;
    postalCode?: string;
    country!: string;
    longitude: number = 1;
    latitude: number = 1;
    phone?: string;
    websiteUrl?: string;
    brewerId?: number;

}